+++ lib/common/data/device_api.dart
