+++ lib/features/memory_recall_training/presentation/screens/memory_recall_chat_screen.dart
@@ -1,4 +1,4 @@
-import 'package:deepscent_cnu/features/normal_olfactory_training/presentation/screens/memory_recall_result_screen.dart';
+import 'package:deepscent_cnu/features/memory_recall_training/presentation/screens/memory_recall_result_screen.dart';
 import 'package:flutter/material.dart';
 
 class MemoryRecallChatScreen extends StatefulWidget {
